Splitting out style hooks
